+++ src/constants/MoveHistory/EventTemplates/index.js
@@ -63,4 +63,4 @@ export { default as rejectSITAddressUpdate } from './UpdateSITAddress/RejectSITA
 export { default as createSITAddressUpdate } from './UpdateSITAddress/CreateSITAddressUpdate';
 export { default as updateMTOShipmentPPMDetails } from './UpdateMTOShipment/UpdateMTOShipmentPPMDetails';
 export { default as updateWeightTicket } from './UpdateWeightTicket/UpdateWeightTicket';
-export { default as createMTOShipmentDetails } from './CreateMTOShipment/createMTOShipmentPPM';
+export { default as createMTOShipmentPPMDetails } from './CreateMTOShipment/createMTOShipmentPPMDetails';
